Manage active descendant on menu role
A menu with a filtering input at the top cannot itself be role=menu
because it hides the input from screen readers. The following
relationship is required.

    <details-menu role="none">
      <input>
      <div role="menu" aria-activedescendant="i1">
        <button id="i1" role="menuitem></button>
      </div>
    </details-menu>

Managing focus with aria-activedescendant requires the attribute to be
attached to the role=menu element in order for the focused item to be
announced.