Scroll details menu into view when it opens off-screen
If the details menu button is at the bottom of the page. we don't
automatically scroll the page to show you the dropdown options.

While there is an `autofocus` attribute we can use, there are two
problems with it:
1. it only works with inputs
2. it doesn't scroll smoothly so the experience not as smooth when
you open the menu.

Let's use scrollIntoView and center on the details menu.